refactoring flex into smaller modules
