+++ coveo-functools/coveo_functools/flex/__init__.py
@@ -1,0 +1,5 @@
+# backward compatibility stuff and import shortcuts
+
+from coveo_functools.flex.decorator import flex, RAW_KEY  # noqa: F401
+from coveo_functools.flex.deserializer import deserialize  # noqa: F401
+from coveo_functools.flex.types import *  # noqa: F401,F403
